LayoutTests/fast/dom/Geolocation/not-enough-arguments.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<script>
description("Tests how Geolocation handles receiving too few arguments.");

shouldThrow("window.navigator.geolocation.getCurrentPosition()");
shouldThrow("window.navigator.geolocation.watchPosition()");
shouldThrow("window.navigator.geolocation.clearWatch()");
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
